requirements.txt
google-genai>=1.0
openai>=1.0
pydantic>=2.0
langgraph>=0.2
streamlit>=1.35
python-dotenv>=1.0
